Rename "capstan" package as "core"
The "go build" command won't work if there's a "capstan" directory in
the top-level directory. Rename the package to "core" instead.
